### 2019-10-07 #1788 @hbaniecki
Hi @terrytangyuan, just by looking at `Domains` in your document i can see that @expectopatronum has written `Interpretable machine learning` there, which is cool.
I have also intersected all of your 723 reviewers with 481 stargazers of DALEX R package. This has led me to 7 people. Some of them list `R` as their language and `Machine Learning` as their domain: @mkearney, @alexpghayes, @acolum .
