### 2019-10-08 #1788 @expectopatronum
I haven't worked with R for about a year now (switched jobs) but my main focus is interpretable machine learning. If you don't find anyone more suitable I'd gladly review this submission.
Off topic: Does anyone know where I can update my entry in this spreadsheet?
