### 2019-10-08 #1788 @alexpghayes
I don't have the bandwidth to review at the moment, but this looks like a cool project!
