### 2019-10-10 #1798 @whedon
[REVIEW]: modelStudio: Interactive Studio with Explanations for ML Predictive Models
**Submitting author:** @hbaniecki (<a href="http://orcid.org/0000-0001-6661-5364">Hubert Baniecki</a>)
**Repository:** <a href="https://github.com/ModelOriented/modelStudio" target ="_blank">https://github.com/ModelOriented/modelStudio</a>
**Version:** v0.1.9
