### 2019-10-10 #1798 @hbaniecki
Great! I am aware that i need to fix paper.bib file to properly code references. I will probably shorten the example. Then Figure 1 should fit one page earlier (same with Figure 2).
