### 2019-10-18 #1798 @acolum
Everything in the software and documentation looks good, but in the paper, there's some spelling, grammar, and citation errors.
