### 2019-10-19 81ef509
fix modelStudio on smaller devices
