### 2019-10-19 62c19b1
update examples and lower B
