+++ DESCRIPTION
@@ -15,7 +15,7 @@ Description: Automate explanation of machine learning predictive models.
   a fast and condensed way to get all the answers without much effort. 
   Break down your model and look into its ingredients with only a few lines of code.
 Depends: R (>= 3.5.0)
-License: GPL
+License: file LICENSE
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
