+++ misc/paper.bib
@@ -10,7 +10,7 @@
   year      = {2016},
 }
 
-@article{2018arXiv180608915B,
+@article{DALEX,
 	author = {{Biecek}, Przemysław},
     title = "{DALEX: explainers for complex predictive models}",
     url = {http://arxiv.org/abs/1806.08915},
@@ -197,9 +197,11 @@ archivePrefix = {arXiv},
   year={2016}
 }
 
-@Manual{pmmeed,
+@Manual{pmeed,
     title = {Predictive Models: Explore, Explain, and Debug},
     author = {Przemyslaw Biecek and Tomasz Burzykowski},
     year = {2019},
+    note = {https://pbiecek.github.io/PM_VEE/},
     url = {https://pbiecek.github.io/PM_VEE/}
 }
+
