### 2019-10-19 2885cc1
Update paper.md
