### 2019-10-20 #1798 @acolum
The references and citations look great, but I'm still noticing a few spelling errors in the first paragraph under "Introduction" and the paragraph under "Conclusions."
