### 2019-10-21 #1798 @expectopatronum
Yes, I also found several typos / missing articles. I marked them in the PDF: [10.21105.joss.01798_typos.pdf](https://github.com/openjournals/joss-reviews/files/3749652/10.21105.joss.01798_typos.pdf)
