+++ misc/paper.md
@@ -23,15 +23,15 @@ affiliations:
 # Introduction
 
 Machine learning predictive models are widely used in many areas of
-bussiness and research. Their rising popularity is due to them beeing
+business and research. Their rising popularity is due to them beeing
 effective but often lead to problems with explaining their prediction.
 This has led to development of many Interpretable Machine Learning
-tools, e.g. `DALEX` [@DALEX] R package, `lime` [@lime] and
+tools, e.g. `DALEX` [@DALEX] R package, `lime` [@lime] and
 `shap` [@NIPS2017_7062] Python packages and `H2o.ai Driverless AI` [@Gill2017MachineLI].
 
-Nowadays, we can see huge demand for automation in many areas. This is
+Nowadays, we can see a huge demand for automation in many areas. This is
 how Automated Machine Learning and Automated Exploratory Data Analysis
-came to existance. AutoML [@Truong2019TowardsAM] and AutoEDA [@2018arXiv180401955S]
+came to existence. AutoML [@Truong2019TowardsAM] and AutoEDA [@2018arXiv180401955S]
 tools not only speed up the model development process
 but also often lead to new discoveries or higher quality of models.
 
@@ -43,9 +43,9 @@ compares various models and explains them. But such tools generate static report
 
 # The `modelStudio` package
 
-The `modelStudio` R package automates the process of model exploration. It generates advanced interactive and animated model explanations in the form of serverless HTML site. It combines **R** [@R] with **D3.js** [@bostock2016d3] to produce plots and descriptions for various local and global explanations. Tools for model exploration unite with tools for EDA to give broad overview of the model behaviour.
+The `modelStudio` R package automates the process of model exploration. It generates advanced interactive and animated model explanations in the form of a serverless HTML site. It combines **R** [@R] with **D3.js** [@bostock2016d3] to produce plots and descriptions for various local and global explanations. Tools for model exploration unite with tools for EDA to give a broad overview of the model behaviour.
 
-The usage of `modelStudio` is meant to be intuitive and simple to use. The computation time needed to produce the output might not be short though.
+The usage of `modelStudio` is meant to be intuitive and simple. The computation time needed to produce the output might not be short though.
 The main goal of this tool is to make model explaining more automated and achieve higher quality explanations by juxtaposition of complementary aspects of a model.
 
 Comparing instance level explanations and model level
@@ -85,10 +85,10 @@ modelStudio(explainer, new_observations)
 
 The generated HTML site has many interactive features. One can choose
 which plots are displayed on the grid and change them at any given
-moment by clicking the X symbol. Drop down list may be used to pick the
-observation that will be considered for local explanation plots. One may manipulate plots having a variable-based dimension by selecting corresponding bars on the other plots. Mousing over D symbol next to explanations displays text descriptions. Finally, mousing over lines and bars displays the tooltip.
+moment by clicking the X symbol. A drop down list may be used to pick the
+observation that will be considered for local explanation plots. One may manipulate plots having a variable-based dimension by selecting corresponding bars on the other plots. Mousing over the D symbol displays a description of the plot. Finally, mousing over lines and bars displays the tooltip.
 
-![1. Open in browser or save as HTML document or PNG image 2. Click on bars to choose which feature will be used for other plots 3. Mouse over D to display plot description and click X to close the plot 4. Choose which observation will be used for local explanations 5. Mouse over lines and bars to display the tooltip 6. Click on the text to choose the plot 7. Interact with other elements like slider](images/cheatsheetcut.png)
[…]
