### 2019-10-22 #1798 @hbaniecki
Hopefully it is clean now.
