### 2019-10-23 ModelOriented/modelStudio#28 @expectopatronum
Missing parts in documentation
Hi, I am one of the [reviewers for your JOSS submission](https://github.com/openjournals/joss-reviews/issues/1798). I thought I'd put the things I miss in the documentation and the corresponding review checklist items here:
* **A statement of need:** It is described what the software should solve, but I somehow miss what the target audience is. Is it researcher, machine learning practitioners, anyone interested in interpretable machine learning...?
* **Installation instructions:** (this might be because I haven't used R much in the past year - as stated before I started the review). When I installed your package (on a Manjaro machine), I had issues because it also installed `glmnet` for which it requires `gcc-fortran` (which I had to install using my package manager). First I am wondering why it knew that it had to install glmnet - it is not mentioned in this libraries DESCRIPTION (I assume it is a dependency of one of the other packages?) And I am also not sure if it is required that your README mentions that one might need to install `gcc-fortran` (because it is not directly used by your package). Just wanted to let you know that this might be an issue :)
* **Automated tests:** The reviewing check list asks "Are there automated tests or manual steps described so that the functionality of the software can be verified?" I can't find such a thing, maybe you can point me to it.
*  **Community guidelines:** The reviewing check list asks "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" I can't find such a thing, maybe you can point me to it.
