### 2019-10-23 ModelOriented/dime#28 @hbaniecki
Hi, I am glad to see this checklist! Here are some notes:

1.  I think that there is no specific target for this package (so basicaly anyone interested in explaining their model). The `modelStudio` package is a part of the [DrWhy.AI](https://github.com/ModelOriented/DrWhy/blob/master/README.md) universe (it is stated in `README.md`) and more information can be found there.
2. This package is on CRAN. Installation should be easy and instructions are clear (in `README.md`). It automatically installs dependencies (other R packages that are stated in Depends/Imports in the `DESCRIPTION` file with all of their dependencies and so on). The `glmnet` package is an import for `ingredients` package (very heavy though). I will try to remove it from Imports in the `ingredients` package. `glmnet` usage is in development stage and it is not needed for the `modelStudio` package.
3. This package has a 99% test coverage and all of the tests are automated. They can be found in `tests` folder. Continuous integration is made possible by [Travis CI](https://travis-ci.org/). Basically it will test code on every commit/pull request on this repository. This package is also tested by CRAN and complies with their (for some people strict) policies. Results of this checks (for different OS) can be found [here](https://cran.r-project.org/web/checks/check_results_modelStudio.html).
4.  Sure, I will add `CONTRIBUTING.md` file to the repo.

I remain at your disposal for any further questions :smile:
