### 2019-10-24 804940e
Create CONTRIBUTING.md
[JOSS submission review](https://github.com/openjournals/joss-reviews/issues/1798)
[modelStudio issue](https://github.com/ModelOriented/modelStudio/issues/28)
