+++ CONTRIBUTING.md
@@ -1,0 +1,34 @@
+## Contribute to modelStudio
+
+#### **Did you find a bug?**
+
+Please follow these rules when reporting bugs:
+
+* Install the latest version of [modelStudio from GitHub](https://github.com/ModelOriented/modelStudio) and check whether the problem still occurs.
+* Install the latest version of the main dependencies [DALEX from Github](https://github.com/ModelOriented/DALEX),
+[ingredients from GitHub](https://github.com/ModelOriented/ingredients) and
+[iBreakDown from GitHub](https://github.com/ModelOriented/iBreakDown).
+
+* **Check, if the bug was not already reported** by searching through [Issues](https://github.com/ModelOriented/modelStudio/issues).
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/ModelOriented/modelStudio/issues/new).
+Be sure to include a **title**, a clear **description** and a **sample code** demonstrating the problem.
+
+
+#### **Did you fix a bug?**
+
+* Ensure that changes meet the requirements of [The tidyverse style guide](http://style.tidyverse.org) and relevant tests are added.
+
+* Open a new GitHub pull request with the solution.
+
+* Ensure that the PR clearly describes the problem and solution. Include the relevant issue number if applicable.
+
+
+#### **Do you intend to add a new feature or change an existing one?**
+
+* Suggest your change in the [Issues](https://github.com/ModelOriented/modelStudio/issues) and start writing your code :smile:.
+
+
+Thanks!
+
+Hubert Baniecki and the ModelOriented team
