### 2019-10-24 ModelOriented/dime#28 @expectopatronum
Thanks, this resolves my questions/concerns!
