### 2019-10-24 #1798 @hbaniecki
Hi! I would like to answer any further questions or resolve remarks to finish the review process (hopefully, by the end of the week) . 🔜 🔚 @acolum @expectopatronum
