### 2019-10-24 #1798 @acolum
@hbaniecki Everything looks great, but there's still a few typos in the paper. In the second sentence under "Introduction," can you correct the spelling of "beeing" to "being" and "lead" to "leads"? After these are fixed, I'd be happy to approve everything for publication.
