### 2019-10-25 #1798 @expectopatronum
Same here, after the typos are fixed, I can accept the paper.
