### 2019-10-25 #1798 @acolum
Looks great to me! I'm finished with my review and can now recommend this package and paper for publication.
