### 2019-11-04 #1798 @hbaniecki
well, it will be `v0.1.9` and `10.5281/zenodo.3527770`
