### 2019-11-04 #1798 @terrytangyuan
@hbaniecki Could you fix the missing DOIs listed above?
