+++ misc/paper.bib
@@ -8,6 +8,7 @@
                13-17, 2016},
   pages     = {1135--1144},
   year      = {2016},
+  doi = {10.18653/v1/n16-3020}
 }
 
 @article{DALEX,
@@ -48,7 +49,8 @@ url = {http://papers.nips.cc/paper/7062-a-unified-approach-to-interpreting-model
      year = 2018,
 	url = {http://arxiv.org/abs/1804.01955},
    adsurl = {http://adsabs.harvard.edu/abs/2018arXiv180401955S},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+  doi = {10.32614/rj-2018-072}
 }
 
 @article{Truong2019TowardsAM,
@@ -67,7 +69,7 @@ url = {http://papers.nips.cc/paper/7062-a-unified-approach-to-interpreting-model
     year = {2019},
     url = {https://www.R-project.org/},
   }
-  
+
 @article{bostock2016d3,
   title={D3.js-data-driven documents (2016)},
   author={Bostock, Mike},
@@ -121,7 +123,8 @@ https://github.com/ModelOriented/ingredients},
     volume = {3},
     number = {26},
     page = {786},
-    url = {https://doi.org/10.21105/joss.00786}
+    url = {https://doi.org/10.21105/joss.00786},
+    doi = {10.21105/joss.00786}
 }
 
 @Manual{tfexplain,
@@ -187,7 +190,8 @@ archivePrefix = {arXiv},
   url = {http://journal.r-project.org/archive/2017/RJ-2017-016/index.html},
   pages = {421--436},
   volume = {9},
-  number = {1}
+  number = {1},
+  doi = {10.32614/rj-2017-016}
 }
 
 @article{apley2016visualizing,
