### 2019-11-04 #1798 @hbaniecki
@terrytangyuan I have added missing DOIs to the .bib file.
