### 2019-11-04 #1798 @terrytangyuan
This paper looks good to me now! Handing over to @openjournals/joss-eics!
