### 2019-11-05 c227ac9
Edits to paper
