### 2019-11-05 #1798 @kyleniemeyer
Hi @hbaniecki, I made a few edits to the paper and references, if you could merge these:
https://github.com/ModelOriented/modelStudio/pull/31 and https://github.com/ModelOriented/modelStudio/pull/32
