### 2019-11-05 #1798 @hbaniecki
Hi @kyleniemeyer, I merged your PR.
