### 2019-11-05 #1798 @kyleniemeyer
@hbaniecki sorry, one last fix I noticed when reviewing: https://github.com/ModelOriented/modelStudio/pull/33
