### 2019-11-05 #1798 @hbaniecki
@kyleniemeyer np. I merged.
