### 2019-11-05 #1798 @kyleniemeyer
Congrats @hbaniecki on your article's publication in JOSS! Many thanks to @acolum and  @expectopatronum for reviewing this, and @terrytangyuan for editing.
