### 2019-11-05 #1798 @hbaniecki
I feel ashamed but hope that it can be updated. Fixed `paper.md`.
