### 2019-11-05 #1798 @arfon
> Oh no! Um, I think so. Can you fix the source file and let us know? (@arfon?)

No problem. Paper updated based on the latest `paper.md` in `master`. FWIW, the new PDF might not show up for a few hours as there's caching in place for the PDFs on the JOSS site.
